requirements.txt
python-craigslist
twilio
pyshorteners
pprint